pyproject.toml
[tool.poetry]
name = "materialcard"
version = "0.1.0"
description = ""
authors = []
packages = [{ include = "materialcard", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
typer = { extras = ["all"], version = "^0.12.0" }
pydantic = "^2.6.0"
pypdf = "^4.0.0"
docxtpl = { version = "^0.16.7", optional = true }
pyyaml = { version = "^6.0.1", optional = true }

[tool.poetry.extras]
docx = ["docxtpl"]
yaml = ["pyyaml"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.3.0"

[tool.poetry.scripts]
materialcard = "materialcard.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
fix = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

